- add support to exclude merge commits in release notes generation - move filterCommits into commits.ts - recompile dist files with updated code
